Make layers as a python module (#6564)
* Make cast op support bool

Also add `elemwise_sub/mul/abs/clip` layers

* Make fuild.layers as a module

* Move layers as a module

* Split layers.py into layers module

* Fix CI

* Fix CI
